requirements.txt
pypdf
langchain
google-generativeAI
streamlit
python-dotenv
